Add failing tests in test file, move processing of exhaustive search into new file
