+++ process_exhaustive_search.py
@@ -1,0 +1,23 @@
+import numpy as np
+
+
+def get_minimum_fofc(csv_name):
+    data = np.genfromtxt('{}.csv'.format(csv_name), delimiter=',', skip_header=0)
+
+    # If four column data from multiple ligand
+    if len(data[0]) == 4:
+        occ = data[:, 0]
+        u_iso = data[:, 2]
+        fo_fc = data[:, 3]
+    elif len(data[0]) == 3:
+        occ = data[:, 0]
+        u_iso = data[:, 1]
+        fo_fc = data[:, 2]
+    else:
+        print("Data is not in correct format")
+    # b_iso = (8 * np.pi ** 2) * u_iso ** 2
+
+    # if three column data
+
+    min_index = np.argmin(fo_fc)
+    return occ[min_index], u_iso[min_index]
