+++ test_exhaustive_search.py
@@ -1,0 +1,47 @@
+import unittest
+
+
+# TODO Write TestComputeMaps
+class TestComputeMaps(unittest.TestCase):
+    """
+    Test the ability to compute maps using compute_maps()
+    """
+
+    def test_compute_maps(self):
+        self.assertEqual(True, False)
+
+
+# TODO Write TestExhaustiveSearch
+class TestExhaustiveSearch(unittest.TestCase):
+    """
+    Test the main loop of exhaustive search.
+    
+    What mock objects does this need?
+    """
+
+    def test_exhaustive_search(self):
+        self.assertEqual(True, False)
+
+
+class TestOccupancyCartPoints(unittest.TestCase):
+    """
+    Test the generation of cartesian points to loop over
+
+    What mock objects does this need?
+    """
+
+    def test_occupancy_cart_points(self):
+        self.assertEqual(True, False)
+
+
+class TestCalculateMeanFofcCartSites(unittest.TestCase):
+    """
+    Test the calculation of |Fo-Fc| given cartesian sitess
+    """
+
+    def test_calc_mean_fofc_cart_sites(self):
+        self.assertEqual(True, False)
+
+
+if __name__ == '__main__':
+    unittest.main()
